working on the beam stuff
based commit
