+++ Assets/Programming/Enemies/Common AI/Ranged/AICommonBeam.cs
@@ -12,27 +12,7 @@ using UnityEngine;
 //  \__,_|\___/|_| |_| |_|_|_.__/|_|  \___/|_| |_|
 // © 2025 Dominic McNeill dommcneill@outlook.com
 
-[Serializable]
-public class BeamAttackSettings
-{
-    public float coolDown = 5f;
 
-    public float tickDamage = 5f;
-
-    public float tickRate = 0.5f;
-
-    public float windUpTime = 1f;
-
-    public float attackDuration = 2f;
-
-    public float beamRadius = 1f;
-
-    public float beamMaxRange = 999f;
-
-    public float turnSpeedWhileCharging = 1f;
-
-    public LineRenderer lineRenderer;
-}
 
 
 public class AICommonBeam : AIBase
