+++ Assets/Programming/Enemies/Common AI/Ranged/BeamAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b125c67a07feb614c8f98c7b9ad54300
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
